+++ madgraph/iolibs/export_v4.py
@@ -3874,6 +3874,8 @@ class ProcessExporterFortranME(ProcessExporterFortran):
         self.copy_python_file()
         
         if model["running_elements"]:
+            if not os.path.exists(pjoin(MG5DIR, 'Template',"RUNNING")):
+                raise Exception("Library for the running have not been installed. To install them please run \"install RunningCoupling\"")
             misc.copytree(pjoin(MG5DIR, 'Template',"Running"), 
                             pjoin(self.dir_path,'Source','RUNNING'))
